backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20
easyocr==1.7.2
Pillow==12.0.0
torch==2.9.1
torchvision==0.24.1
transformers==4.57.1
deep-translator==1.11.4
pyttsx3==2.99
numpy==2.2.6
opencv-python-headless==4.12.0.88
requests==2.32.5
